Move from forking systemd service to simple
This way, the service is not recognised as "active" if the DHCP server is not actually running.
However, if the DHCP server IS actually running, the service is also not recognised as "active", though the status is success.
Checking if the DHCP server is actually running requires probing for isc-dhcp-server directly.